Like button bug fix.
Fixed bug with page rendering when pressing like button. The bug was that when the like button is clicked and the page is reloaded, the post author is not shown because when liking, post is being found by it's ID but the author is not populated to the found post. The bug was fixed by adding ".populate('author')" when searching the current post in the database.